added CardDataSource methods to QuizDataSource. Going to try and CardDataSource refactor next to see if it work. if not will just reset back to here
